refactor(ads): ネイティブ広告の表示方式を改善
- NativeAdCacheクラスを削除し、シンプルなマップベース保持に変更
- 一度読み込んだ広告を位置ごとに保持し、再表示時の再読み込みを防止
- 画面外にスクロールした後の再表示でも同じ広告を表示
- 複雑なキャッシュロジックを排除し、メモリ効率と安定性を向上
