+++ composeApp/src/androidMain/kotlin/com/maropiyo/reminderparrot/ui/components/common/ad/AndroidAdFactory.kt
@@ -14,22 +14,26 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.viewinterop.AndroidView
+import com.google.android.gms.ads.AdListener
+import com.google.android.gms.ads.AdLoader
 import com.google.android.gms.ads.AdRequest
 import com.google.android.gms.ads.AdSize
 import com.google.android.gms.ads.AdView
+import com.google.android.gms.ads.LoadAdError
 import com.google.android.gms.ads.nativead.NativeAd
+import com.google.android.gms.ads.nativead.NativeAdOptions
 import com.google.android.gms.ads.nativead.NativeAdView
 import com.maropiyo.reminderparrot.domain.usecase.GetUserSettingsUseCase
 
 class AndroidAdFactory(
     private val getUserSettingsUseCase: GetUserSettingsUseCase
 ) : AdFactory {
-    private var nativeAdCache: NativeAdCache? = null
+    // 読み込み済み広告を保持するマップ
+    private val loadedAds = mutableMapOf<Int, NativeAd>()
 
     @Composable
     override fun BannerAd(modifier: Modifier) {
@@ -75,49 +79,36 @@ class AndroidAdFactory(
         }
 
         val context = LocalContext.current
-        val scope = rememberCoroutineScope()
-        var nativeAd by remember(adPosition) { mutableStateOf<NativeAd?>(null) }
+        var nativeAd by remember(adPosition) { mutableStateOf<NativeAd?>(loadedAds[adPosition]) }
 
-        // キャッシュを初期化
-        if (nativeAdCache == null) {
-            nativeAdCache = NativeAdCache(context, scope)
-        }
-
-        // キャッシュから広告を取得または事前読み込み
+        // 広告をロード（まだ読み込んでいない場合のみ）
         LaunchedEffect(adPosition) {
-            val cache = nativeAdCache ?: return@LaunchedEffect
+            if (loadedAds[adPosition] == null) {
+                val adLoader = AdLoader.Builder(context, "ca-app-pub-3940256099942544/2247696110")
+                    .forNativeAd { ad ->
+                        loadedAds[adPosition] = ad
+                        nativeAd = ad
+                    }
+                    .withAdListener(object : AdListener() {
+                        override fun onAdFailedToLoad(adError: LoadAdError) {
+                            println("ネイティブ広告の読み込みに失敗: ${adError.message}")
+                        }
+                    })
+                    .withNativeAdOptions(
+                        NativeAdOptions.Builder()
+                            .setAdChoicesPlacement(NativeAdOptions.ADCHOICES_TOP_RIGHT)
+                            .build()
+                    )
+                    .build()
 
-            // キャッシュされた広告があるかチェック
-            val cachedAd = cache.getAd(adPosition)
-            if (cachedAd != null) {
-                nativeAd = cachedAd
-                println("📱 AndroidAdFactory: ✅ キャッシュから広告を取得 (position: $adPosition)")
-            } else {
-                println("📱 AndroidAdFactory: ❌ キャッシュなし、事前読み込み開始 (position: $adPosition)")
-                // キャッシュにない場合は事前読み込み
-                cache.preloadAd(adPosition)
-
-                // 少し待ってから再度チェック
-                kotlinx.coroutines.delay(1500)
-                val newAd = cache.getAd(adPosition)
-                if (newAd != null) {
-                    nativeAd = newAd
-                    println("📱 AndroidAdFactory: ✅ 事前読み込み完了 (position: $adPosition)")
-                } else {
-                    println("📱 AndroidAdFactory: ⚠️ 事前読み込み失敗、ダミー表示 (position: $adPosition)")
-                    // フォールバック: ダミーデータで表示
-                    // 実際の広告がない場合でも何かしら表示するためのダミー広告
-                }
+                adLoader.loadAd(AdRequest.Builder().build())
             }
-
-            // 次の広告も事前読み込み
-            cache.preloadAds(listOf(adPosition + 5, adPosition + 10))
         }
 
         AndroidView(
             modifier = modifier,
             factory = { context ->
-                createNativeAdView(context, null)
+                createNativeAdView(context)
             },
             update = { view ->
                 (view as? NativeAdView)?.let { adView ->
@@ -151,7 +142,7 @@ class AndroidAdFactory(
         )
     }
 
-    private fun createNativeAdView(context: Context, nativeAd: NativeAd?): NativeAdView {
+    private fun createNativeAdView(context: Context): NativeAdView {
         val adView = NativeAdView(context)
 
         // メインコンテナ
